Place decorative borders around panes in dual-pane mode.
Introduce new class NautilusBorder, which is a border (frame) widget
that knows an active and an inactive state. When active, the border is
drawn in the theme's "selected" color. When inactive, it looks like a
sunk-in shadow.

When a window is put in dual-pane mode, both sides are placed in such
a border.