Remember extra pane status access windows.
The status of the extra pane display is remembered in a gconf key,
analog to show/hide statusbar and other window display properties. As
the extra-pane action implementation does a lot of widget and action
re-ordering, this can only be done when a windows is set up
completely, and is thus delayed in an idle-callback. If splitting the
view is to be done on startup, the showing of the window is delayed to
avoid ugly visible reordering.